Removed format-specific transform-ready->absolute/relative utils
refs https://github.com/TryGhost/Team/issues/467

- transform-ready to absolute/relative is a straightforward regex replace on a string so there is no need to have format-specific utility functions
- moved all format-specific tests into the `transformReadyToAbsolute/Relative` tests to cover any possible future regressions
