--- packages/url-utils/lib/utils/html-transform-ready-to-absolute.js
@@ -1,20 +1,0 @@
-function escapeRegExp(string) {
-    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-}
-
-const htmlTransformReadyToAbsolute = function (html = '', root, _options = {}) {
-    const defaultOptions = {
-        replacementStr: '__GHOST_URL__'
-    };
-    const options = Object.assign({}, defaultOptions, _options);
-
-    if (!html || html.indexOf(options.replacementStr) === -1) {
-        return html;
-    }
-
-    const replacementRegex = new RegExp(escapeRegExp(options.replacementStr), 'g');
-
-    return html.replace(replacementRegex, root.replace(/\/$/, ''));
-};
-
-module.exports = htmlTransformReadyToAbsolute;
