+++ packages/url-utils/lib/utils/index.js
@@ -9,21 +9,15 @@ module.exports = {
     htmlRelativeToAbsolute: require('./html-relative-to-absolute'),
     htmlAbsoluteToTransformReady: require('./html-absolute-to-transform-ready'),
     htmlRelativeToTransformReady: require('./html-relative-to-transform-ready'),
-    htmlTransformTreadyToAbsolute: require('./html-transform-ready-to-absolute'),
-    htmlTransformTreadyToRelative: require('./html-transform-ready-to-relative'),
     isSSL: require('./is-ssl'),
     markdownAbsoluteToRelative: require('./markdown-absolute-to-relative'),
     markdownRelativeToAbsolute: require('./markdown-relative-to-absolute'),
     markdownAbsoluteToTransformReady: require('./markdown-absolute-to-transform-ready'),
     markdownRelativeToTransformReady: require('./markdown-relative-to-transform-ready'),
-    markdownTransformTreadyToAbsolute: require('./markdown-transform-ready-to-absolute'),
-    markdownTransformTreadyToRelative: require('./markdown-transform-ready-to-relative'),
     mobiledocAbsoluteToRelative: require('./mobiledoc-absolute-to-relative'),
     mobiledocRelativeToAbsolute: require('./mobiledoc-relative-to-absolute'),
     mobiledocAbsoluteToTransformReady: require('./mobiledoc-absolute-to-transform-ready'),
     mobiledocRelativeToTransformReady: require('./mobiledoc-relative-to-transform-ready'),
-    mobiledocTransformTreadyToAbsolute: require('./mobiledoc-transform-ready-to-absolute'),
-    mobiledocTransformTreadyToRelative: require('./mobiledoc-transform-ready-to-relative'),
     relativeToAbsolute: require('./relative-to-absolute'),
     relativeToTransformReady: require('./relative-to-transform-ready'),
     replacePermalink: require('./replace-permalink'),
