+++ packages/url-utils/lib/utils/transform-ready-to-absolute.js
@@ -1,20 +1,20 @@
-const deduplicateSubdirectory = require('./deduplicate-subdirectory');
+function escapeRegExp(string) {
+    return string.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+}
 
-const transformReadyToAbsolute = function transformReadyToAbsolute(url, root, _options = {}) {
+const transformReadyToAbsolute = function (str = '', root, _options = {}) {
     const defaultOptions = {
         replacementStr: '__GHOST_URL__'
     };
     const options = Object.assign({}, defaultOptions, _options);
 
-    if (url.indexOf(options.replacementStr) !== 0) {
-        return url;
+    if (!str || str.indexOf(options.replacementStr) === -1) {
+        return str;
     }
 
-    const transformedUrl = url
-        .replace(options.replacementStr, root)
-        .replace(/([^:])\/\//g, '$1/');
+    const replacementRegex = new RegExp(escapeRegExp(options.replacementStr), 'g');
 
-    return deduplicateSubdirectory(transformedUrl, root);
+    return str.replace(replacementRegex, root.replace(/\/$/, ''));
 };
 
 module.exports = transformReadyToAbsolute;
